Add a option to use the module behind a proxy.
Some people host their Django applications beyond a set of proxies, which
mean that REMOTE_ADDR is the ip address of the proxies, not of the client.
Most of them usually place the original IP address in a different HTTP header,
so making this configurable make stopforumspam work in such a environment.